pyproject.toml
[project]
name = "Magic_Eye_Generator"
version = "0.1.0"
description = "Magic Eye Image Generator with adaptive CUDA support. It also includes training features for game difficulty estimation."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.133.1",
    "fastapi-utils>=0.8.0",
    "firebase-admin>=7.2.0",
    "googletrans>=4.0.2",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "numpy>=2.0.0",
    "pillow>=10.0.0",
    "uvicorn>=0.30.0",
    "accelerate>=0.26.0",
    "diffusers>=0.26.0",
    "opencv-python>=4.8.0",
    "torch>=2.1.0,<2.6.0",
    "torchvision>=0.16.0,<0.21.0",
    "transformers>=4.35.0",
    "google-cloud-storage>=2.14.0",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "ruff>=0.1.0"]

[tool.uv]
environments = []

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
# 윈도우 환경(win32)에서만 pytorch-cuda 인덱스를 사용하도록 지정합니다.
# 다른 환경에서는 기본 인덱스(PyPI)를 자동으로 사용합니다.
torch = [
  { index = "pytorch-cuda", marker = "sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cuda", marker = "sys_platform == 'win32'" },
]

[tool.ruff]
ignore = ["E501"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]